+++ modules/apps/27-interchain-accounts/genesis.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/ibc-go/modules/apps/ibc-account/keeper"
-	"github.com/cosmos/ibc-go/modules/apps/ibc-account/types"
+	"github.com/cosmos/ibc-go/modules/apps/27-interchain-accounts/keeper"
+	"github.com/cosmos/ibc-go/modules/apps/27-interchain-accounts/types"
 )
 
 func InitGenesis(ctx sdk.Context, keeper keeper.Keeper, state types.GenesisState) {
